+++ BE/src/controllers/auth.ts
@@ -2,10 +2,11 @@ import createError from 'http-errors';
 import { Context, Next } from 'koa';
 import axios from 'axios';
 
-import service from '@/services/auth';
+import serviceAuthCheck from '@/services/auth/check';
+import serviceLogin from '@/services/auth/login';
 
 const login = async (ctx: Context): Promise<Context['body']> => {
-  const token = await service.login(ctx.request.body);
+  const token = await serviceLogin.login(ctx.request.body);
   console.log('token: ', token);
 
   ctx.body = { token };
@@ -15,7 +16,7 @@ const checkToken = async (
   ctx: Context,
   next: Next,
 ): Promise<Context['body']> => {
-  const isUserInDB = await service.checkToken(ctx.header);
+  const isUserInDB = await serviceAuthCheck.checkToken(ctx.header);
   if (!isUserInDB) {
     const jwtError = createError(401, 'unauthorized');
     throw jwtError;
